#30 refactor: inactive 유저와 없는 유저 구분
