+++ src/main/java/com/floney/floney/user/service/OAuthUserService.java
@@ -6,7 +6,8 @@ import com.floney.floney.user.dto.request.SignupRequest;
 public interface OAuthUserService {
 
     boolean checkIfSignup(String oAuthToken);
-    void signup(String oAuthToken, SignupRequest request);
-    Token login(String oAuthToken);
 
+    void signup(String oAuthToken, SignupRequest request);
+
+    Token login(String oAuthToken);
 }
